Made system stable (#16)
* went over the memory and logic errors with gemini

* made code debug and non debug mode ready

* made system stable and memory efficient

+++ lib/plants/structs/plantConfig.h
@@ -5,7 +5,7 @@ struct PlantConfig
 {
     uint8_t idx;           // Plant idx
     uint8_t sunnyHours;    // Required light hours
-    uint8_t waterLimit;    // At what soil moisture level to water (e.g., 30%)
+    uint16_t waterLimit;   // At what soil moisture level to water (e.g., 30%)
     uint16_t waterMl;      // Watering amount
     int16_t idealTemp;     // Ideal temp (e.g., 22.5C = 225)
     int16_t idealSoilTemp; // Ideal temp (e.g., 22.5C = 225)
